backend/requirements.txt
# Python >= 3.8 required
fastapi==0.109.2
uvicorn==0.27.1
python-dotenv==1.0.1
motor==3.3.2
pymongo==4.6.1
beanie==1.25.0
pydantic==2.6.1
pydantic-settings==2.1.0
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.9
pytest>=8.0.0
httpx>=0.26.0
email-validator==2.1.0.post1
bcrypt==4.1.2